requirements.txt
opencv-python
tqdm
tensorboard
addict
scikit-learn
pathspec
imagesize
ujson
matplotlib
pillow
torchvision
numpy
